Initial commit
Syllable_sequencing is the ipynb->py conversion of the model Senft et al published with their paper in 2062 "Reduction of dopamine in basal ganglia...." and the altered version has three noteworthy changes; progress bar is now terminal instead of HTML, it accepts command line arguments to parameterize tonic dopamine levels represented by lg and le as well as accepting a string as a relative file path to save the produced matlab graph, and the graph itself is now closed to facillitate repeated runs without undue memory leakage.

automator is a simple script to replicate the original study - lg and le are parameterized and run from 0.2 to 0 by steps of 0.02, and each set of parameter values is ran five times. That means 121 sets of parameter values (0.2 to 0 by 0.02 is eleven steps, 11*11=121) and a total of 605 graphs saved. Each is named for the decimal level of dopamine, so a graph titled 0000a is the first run of the lg=le=0 trial, and 2004c is the third run of the lg=0.2, le=0.04 trial.

agreggator is another simple script to take all the graphs from the previous operation and put them in some other directory. The naming convection allows you to then look at each graph sequentially if for example you wanted to count the syllables by hand or get a sense of the overall data.

USEAGE: have the scripts in the same directory. Use ipython magic %matplotlib inline before running the automator script. These are tested to work in the Spyder IDE running Anaconda. The current state of the scripts depends on the runfile() call in automator being modified to reflect your current directory. Simply run any script and the (potentially undocumented) syntax of runfile() will be present in the interactive interpreter.

CURRENT TO DO:
Modify graphs to have consistent color and only save the graphs needed for counting number of syllables sequenced.
Make scripts slightly more robust, at present they depend on literal ipython magic and figurative IDE magic to run. Absolutely no error checking is done.
Save raw data from a given trial for later use, such as automated template matching for number of syllables sequenced
Create script to make 'grading' a trial simpler and multi-user (fancy spreadsheet)