pck.py to experiment with serialized data
serialized data from simulation results, goal of pck is to automate the scoring of an individual simulation by analyzing the raw data. Eventual goal is to keep results of all simulations in pickle format, so as to re-analyze and re-graph any given run. Note that the actual simulation environment is not part of that - would need to verify that the top level network seed number generated as part of the build process is sufficient for deterministic builds every time.